requirements.txt
blinker
certifi
click
dnspython
Flask
itsdangerous
Jinja2
MarkupSafe
pymongo
python-dotenv
Werkzeug
